requirements.txt
pandas
numpy
requests
matplotlib
seaborn
textblob
python-dotenv